Porting attach command(WIP)
